Remove the thing's magic method.
The _no_map magic method is only valid for a few methods in the class, but is callable via __call() on all methods.  This change removes the __call() magic method and adds the required [func]_no_map() methods to the class.

It also replaces a single instance of value_no_map() with value() as _no_map has no effect on the value() function.

This makes the code more readable, solve solves several scrutinizer issues and allows for better doc blocks.