Add hau, bring back ha, change ky to kir
As per @ocean90's request, adding a new locale for Hausa in Latin script and bringing back the data for the old rtl locale. As per a GTE request & information on Ethnologue and Wikipedia, also fixing the Kirghiz locale code (former ky, which is the wrong code for the language) and exchanging the locale codes with the ISO 639-3 code.