Fix newline and tab character filters
Add raw original data to the translations row and use it for the "copy from original" code, then remove the filtering of the invisibles tab/nl so that translations can contain these UTF characters.  Also convert "copy from original" to a button and added two new buttons to make it easier to add tab and newline characters to the translation.