Separate the specific warnings for wporg
Create a new class (GP_Wporg_Translation_Warnings) for the wporg warnings
and inject it using hooks in the GP_Builtin_Translation_Warnings class.

Inject the specific wporg list of domains with allowed changes to their own
subdomains and a specific wporg warning for changing placeholders in the
format of '###[A-Z_]+###'.