Generate .htaccess file when using a wp-env environment
Before this commit, visiting http://localhost:8888/glotpress after following setup instructions would result in a 404. The cause of that 404 was that the .htaccess file had not been generated.

Here we're generating the .htaccess file through the `wp rewrite` command, with the --hard flag:
https://developer.wordpress.org/cli/commands/rewrite/structure/

However, for the `wp rewrite` command to be able to generate/modify .htaccess, the `mod_rewrite` module needs to be enabled in wp-cli's config, so we're adding a new wp-cli.yml file to set that option. This is required because the `apache_modules` setting cannot be passed as a flag to the CLI, it must be passed through a config file.