Add a new locale: Spanish and Guaraní (Paraguay) to locales.php
I am writing to formally request the addition of a new locale for Spanish (Paraguay) and Guaraní (Paraguay) on WordPress. This request is to reopen and follow up on the previously initiated issue, which can be found here: Request for Paraguayan Spanish and Guaraní Locales.

The inclusion of the Paraguayan Spanish and Guaraní locales is essential to cater to the unique linguistic needs of the Paraguayan WordPress community. This locale will address specific regional variations in vocabulary, syntax, and semantics, providing a more localized and relevant user experience for WordPress users in Paraguay.